modify ingredients API
