+++ backend/nutrition_facts/serializers.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from .models import Ingredient
+
+
+class IngredientSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Ingredient
+        fields = '__all__'
